+++ ampltools/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.5.1 - 2023-03-03
+- When a module is not available, try to indicate a bundle with it.
+- Add unit tests for modules.
+
 ## 0.5.0 - 2023-03-02
 - Add modules.activate and activate command.
 - Add modules.preload to load modules silently.